ChzzkPythonProject/requirements.txt
python-socketio
websocket-client
requests
flask